Add UserFavourite#show,index
