--- app/controllers/concerns/UserModule/user_favourite_controller.rb
@@ -1,6 +1,0 @@
-class UserFavouriteController < ApplicationController
-
-
-
-
-end
