--- app/models/UserModule/user.rb
@@ -1,10 +1,0 @@
-class User < ActiveRecord::Base
-
-  has_secure_password
-  validates :email, presence: true, uniqueness: true
-  validates :email, format: { with: URI::MailTo::EMAIL_REGEXP }
-  validates :password,
-            length: { minimum: 6 },
-            if: -> { new_record? || !password.nil? }
-
-end
