--- app/models/UserModule/user_favourite.rb
@@ -1,3 +1,0 @@
-class UserFavourite < ActiveRecord::Base
-
-end
