+++ app/models/user_module/user.rb
@@ -1,0 +1,12 @@
+module UserModule
+  class User < ActiveRecord::Base
+
+    has_secure_password
+    validates :email, presence: true, uniqueness: true
+    validates :email, format: { with: URI::MailTo::EMAIL_REGEXP }
+    validates :password,
+              length: { minimum: 6 },
+              if: -> { new_record? || !password.nil? }
+
+  end
+end
