+++ app/models/user_module/user_favourite.rb
@@ -1,0 +1,5 @@
+module UserModule
+  class UserFavourite < ActiveRecord::Base
+
+  end
+end
