+++ config/routes.rb
@@ -1,14 +1,15 @@
 Rails.application.routes.draw do
 
-  root 'session#home'
-
-  get 'home/index', to: 'home#index'
-
   #User routes
-  resources :user, only: [:new, :create, :index, :show]
+  namespace :user_module do
+    resources :user, only: [:new, :create, :index, :show]
+    resources :user_favourite
+  end
 
+  #Auth routes
   post '/auth/login', to: 'authentication#login'
   get '/*a', to: 'application#not_found'
+
