+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_12_04_151418) do
+ActiveRecord::Schema[7.0].define(version: 2022_12_04_153450) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -21,14 +21,14 @@ ActiveRecord::Schema[7.0].define(version: 2022_12_04_151418) do
     t.datetime "updated_at", null: false
   end
 
-  create_table "user_favourite", force: :cascade do |t|
+  create_table "user_favourites", force: :cascade do |t|
     t.integer "user_id"
     t.integer "topic_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
 
-  create_table "user_read", force: :cascade do |t|
+  create_table "user_reads", force: :cascade do |t|
     t.integer "user_id"
     t.integer "node_id"
     t.datetime "created_at", null: false
